requirements.dev.txt
boto3-stubs ~= 1.35.90
moto ~= 5.0.25
mypy ~= 1.14.1
pandas-stubs ~= 2.2.2.240807
pre-commit ~= 4.0.1
pydoc-markdown ~= 4.8.2
pytest ~= 8.3.4
pytest-dotenv ~= 0.5.2
responses ~= 0.25.6
smart-open ~= 7.1.0
types-defusedxml ~= 0.7.0.20240218
types-requests == 2.31.0.6
